Refactor email service and button to be dynamic based on form route
