+++ app/frontend/src/services/emailService.js
@@ -1,14 +1,14 @@
 import { appAxios } from '@/services/interceptors';
-import { ApiRoutes } from '@/utils/constants';
 
 export default {
   /**
    * @function requestReceiptEmail
    * Sends an application registration request email
-   * @param {object} content An object with ipcPlanId and to attributes
+   * @param {string} form The form name
+   * @param {object} content An object with submissionId and to attributes
    * @returns {Promise} An axios response
    */
-  requestReceiptEmail(content) {
-    return appAxios().post(ApiRoutes.EMAIL, content);
+  requestReceiptEmail(form, content) {
+    return appAxios().post(`${form}/submissions/email`, content);
   }
 };
